+++ CHANGELOG.rst
@@ -1,7 +1,16 @@
 Change Log
 ----------
 
+`Unreleased`_
++++++++++++++
+
+**Added**
+
+- A board data structure
+
 0.0.1 (2016-09-05)
 ++++++++++++++++++
 
 Birth!
+
+.. _`Unreleased`: https://github.com/dwayne/xo-python/compare/v0.0.1...HEAD